Add create / edit game page
